Factor tab panel into single file, delete unnecessary files.
